backend/requirements.txt
arrow==1.3.0
asgiref==3.8.1
blessed==1.20.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.0.7
django-compat==1.0.15
django-cors-headers==4.4.0
django-picklefield==3.2
django-q==1.3.9
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
filelock==3.15.4
idna==3.7
mysqlclient==2.2.4
oauthlib==3.2.2
pipenv==2024.0.1
platformdirs==4.2.2
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
redis==3.5.3
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==72.1.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
types-python-dateutil==2.9.0.20240316
urllib3==2.2.2
virtualenv==20.26.3
wcwidth==0.2.13
wheel==0.43.0
